requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.0.4
django-allauth==0.48.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
fontawesome-free==5.15.3
gunicorn==19.0.0
idna==3.2
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
pypng==0.0.20
PyQRCode==1.2.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0